Distance Calculation from Origin
A robot moves in a plane starting from the original point (0,0). The robot can move toward UP, DOWN, LEFT and RIGHT with a given steps. The trace of robot movement is shown as the following:
UP 5
DOWN 3
LEFT 3
RIGHT 2

The numbers after the direction are steps. The program computes the distance from current position after a sequence of movement and original point. If the distance is a float, then it is rounded to the nearest integer.
Example:
If the following tuples are given as input to the program:
UP 5
DOWN 2
LEFT 6
RIGHT 2
Then, the output of the program should be:
5